# #12122 Speed up Deferred by removing assertions