--- a/src/twisted/newsfragments/12122.feature
+++ b/src/twisted/newsfragments/12122.feature
@@ -1,0 +1,1 @@
+``twisted.internet.defer.Deferred`` operations use less CPU.