## Scope and purpose

Fixes #12122


This meaningfully impacts the HTTP client benchmark I've been using. Before:

```
$ hyperfine --warmup 1 "python http_client_benchmark.py"
Benchmark 1: python http_client_benchmark.py
  Time (mean ± σ):      2.942 s ±  0.026 s    [User: 2.931 s, System: 0.011 s]
  Range (min … max):    2.902 s …  2.989 s    10 runs
```

After:

```
$ hyperfine --warmup 1 "python http_client_benchmark.py"
Benchmark 1: python http_client_benchmark.py
  Time (mean ± σ):      2.847 s ±  0.023 s    [User: 2.834 s, System: 0.012 s]
  Range (min … max):    2.807 s …  2.893 s    10 runs
```

## Some reasons not to merge this!

1. ~Some of these issues are caught by type checking, but not all. In particular, the assertion that the result is not a Deferred isn't caught by type checking, I think.~
2. ~Users could just run with PYTHONOPTIMIZE=2. In practice, that slows down execution(!), not sure why. And I suspect no one runs this way.~
3. Users who don't use type checking will have harder time.