@adiroiban 2024-03-27
Many thanks for the PR

I don't think `PYTHONOPTIMIZE=2` is an option

I think `PYTHONOPTIMIZE` also removes docstrings, and I think that in Twisted we have some exceptions classes that get the exception message from the class docstring.

That is, Twisted is not supported with `PYTHONOPTIMIZE=2`

----------

```diff
-       assert not isinstance(result, Deferred)
        self._startRunCallbacks(result)
```

For the change over here. I don't know what to say.
I don't remember trying to pass a deferred to `Deferred.callback()`, but I might have done it at the start of using Twisted. I can image some new developers might try to do that...and end up with unexpected results


For use cases in which this is an important performance issue, maybe you can do something like:

```
Deferred.callback = Deferred._startRunCallbacks
```

and maybe we can have something like that can be enabled for users at bootstrap:

```
@classmethod
def tuneForPerformance(cls):
    """
    Change the deferred to improve performance.
    Certain safety nets are removed, so you should make sure you code is valid.
    """
    cls.callback = cls._startRunCallbacks
```


we can write a test and make sure this functionality is supported.

In the future, we might add other performance improvements

It might be a stupid idea... but I don't know how else to fix this performance issue in a "standard" way.

So as a general rule...we might have Twisted runnig with "training wheels" by default.
Once you are happy with your code, you can remove the training wheels and expert a small performance improvement.