@glyph 2024-03-27
> I don't remember trying to pass a deferred to `Deferred.callback()`, but I might have done it at the start of using Twisted. I can image some new developers might try to do that...and end up with unexpected results

Previously we had no other way to enforce this. Now, the answer to such confusion would be "you should run mypy on your code". It will catch mismatches like this.