@glyph 2024-03-27
> For use cases in which this is an important performance issue,

Deferreds are in the hot loop of pretty much every Twisted application, so I think we can say that it's OK to assume that this is always a performance issue.

> maybe you can do something like:

Performance should be good by default. Users will not discover a hidden performance tunable like this, so I am -1 on such a change.