@glyph 2024-03-27
I really wish we could have repeatable performance benchmarks in CI though. I miss speed.twistedmatrix.com :-(