@adiroiban 2024-03-27
> Previously we had no other way to enforce this. Now, the answer to such confusion would be "you should run mypy on your code". It will catch mismatches like this.

I don't know if mypy can fail if we pass a Defered instance.
The current signature is just `object`

I read something about `TypeGuard` but I don't know if it can help here... it looks like runtime code  https://mypy.readthedocs.io/en/stable/type_narrowing.html#user-defined-type-guards