@adiroiban 2024-03-27
> I really wish we could have repeatable performance benchmarks in CI though. I miss speed.twistedmatrix.com :-(

With Datadog you can ingess general xunit XML report files and see the execution time for each test... and in theory you can generate a graph with the evolution over time.

I didn't had time to follow up with that... Datadog is not a turn key solution.
But Datadog can provide persistance for test results and has some tools to create graphs / reports / dashboards