@glyph 2024-03-27
`_SelfResultT` is not the same as `object`.