@adiroiban 2024-03-27
I see. Thanks. I wasn't aware of the mypy_zope wizardry :)

This means you need to add type checking to any existing code

```
# test.py
from twisted.internet.defer import Deferred
trigger = Deferred[bytes]()
source = Deferred[int]()
source.callback(trigger)

$ mypy test.py 
test.py:4:17: error: Argument 1 to "callback" of "Deferred" has incompatible type "Deferred[bytes]"; expected "int | Failure"  [arg-type]
Found 1 error in 1 file (checked 1 source file)
```

---------

As long as others are happy with this change, in the name of performance, I am fine with this change.

If needed I can create a custom `add_training_wheels()` functions that adds back type checking.