@glyph 2024-03-27
> If needed I can create a custom `add_training_wheels()` functions that adds back type checking.

There are actually third-party libraries that can translate type hints into assertions, like https://github.com/beartype/beartype .  Not sure how well they work, but I think a general solution is better than a few manually-added hacks like this.