@itamarst 2024-03-27
Actually I may have been wrong and type checking _will_ catch passing a Deferred to `callback()`.

In particular, the signature of `callback()` is `callback(self, result: Union[_SelfResultT, Failure]) -> None:`. And mypy at least forces you to define the generic type of the `Deferred`, which means it will get caught.

Assuming, of course, you're using type checking.