@itamarst 2024-03-27
Oh and apparently this exact discussion happened while I was playing around with it myself, oops.